requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
python-dotenv==1.0.0
